perf(dflash2): store fused-attention query tiles in half registers
The rotated queries originate as fp16. Keeping their scaled tile in half reduces per-SIMDgroup register pressure while retaining fp32 dot products, softmax state, and value accumulation.


Result: {"status":"keep","dflash2_tokens_per_second":37.241407,"dflash2_decode_latency_ms":3410.182666,"dflash2_decode_stddev_ms":16.960765,"dflash2_decode_tokens":127}

+++ crates/mlx-cpp/turbo/turbo4_attention.cpp
@@ -64,7 +64,7 @@ constexpr const char* TURBO4_MTP_VERIFY_PARTIAL_SOURCE = R"(
     float scale_log2e = scale[0] * 1.4426950408889634f;
 
     uint rows[QueriesPerGroup];
-    float q[QueriesPerGroup][DimsPerThread];
+    half q[QueriesPerGroup][DimsPerThread];
     float out[QueriesPerGroup][DimsPerThread];
     float max_score[QueriesPerGroup];
     float sum_score[QueriesPerGroup];
